Refactor ChangeStream to simply an EventEmitter
This decision was made for simplicity sake as well as trying to reduce
the library size by not needing to include all Node Stream modules.  If
one would like to utilize streams to interact with a Nuclear Reactor it
is simply a matter of wrapping the changeEmitter in a writable stream.