Call setState on React components once per dispatch
The current ReactMixin implementation calls `reactor.observe`
separately for each entry in the response to a React component's
getDataBindings. If multiple pieces of bound data change in a single
dispatch, Nuclear will call setState on the component for each of them
and React will dutifully render each time. This means that the exact
behavior of your component is dependent on the order in which you
define your data bindings (aside from being bad for performance).

This patch switches to a single `reactor.observe` call that includes
all of the data bindings for the component.

Fixes #193